Conflicts:
	src/ServiceManagement/StorSimple/Commands.StorSimple/Commands.StorSimple.csproj
	src/ServiceManagement/StorSimple/Commands.StorSimple/ServiceClients/PSStorSimpleVolumeClient.cs

+++ src/ServiceManagement/StorSimple/Commands.StorSimple/Cmdlets/DataContainer/RemoveAzureStorSimpleDeviceVolumeContainer.cs
@@ -5,48 +5,54 @@ using Microsoft.WindowsAzure;
 
 namespace Microsoft.Azure.Commands.StorSimple.Cmdlets
 {
+    using Properties;
+
     [Cmdlet(VerbsCommon.Remove, "AzureStorSimpleDeviceVolumeContainer")]
     public class RemoveAzureStorSimpleDeviceVolumeContainer : StorSimpleCmdletBase
     {
-        [Alias("DN")]
-        [Parameter(Position = 0, Mandatory = true, HelpMessage = "The device name.")]
+        [Parameter(Position = 0, Mandatory = true, HelpMessage = StorSimpleCmdletHelpMessage.HelpMessageDeviceName)]
         [ValidateNotNullOrEmptyAttribute]
         public string DeviceName { get; set; }
 
-        [Alias("DC")]
-        [Parameter(Position = 1, Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = "The volume container name which needs to be removed.")]
+        [Parameter(Position = 1, Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = StorSimpleCmdletHelpMessage.HelpMessageDataContainerName)]
         [ValidateNotNullOrEmptyAttribute]
         public DataContainer VolumeContainer { get; set; }
 
-        [Alias("Wait")]
-        [Parameter(Position = 2, Mandatory = false, HelpMessage = "Wait for remov task complete")]
+        [Parameter(Position = 2, Mandatory = false, HelpMessage = StorSimpleCmdletHelpMessage.HelpMessageWaitTillComplete)]
         public SwitchParameter WaitForComplete { get; set; }
+
+        [Parameter(Position = 3, Mandatory = false, HelpMessage = StorSimpleCmdletHelpMessage.HelpMessageForce)]
+        public SwitchParameter Force { get; set; }
 
         public override void ExecuteCmdlet()
         {
-            try
-            {
-                var deviceid = StorSimpleClient.GetDeviceId(DeviceName);
+            ConfirmAction(Force.IsPresent,
+                          Resources.RemoveWarningVolume,
+                          Resources.RemoveConfirmationVolume,
+                          string.Empty,
+                          () =>
+                          {
+                              try
+                              {
+                                  var deviceid = StorSimpleClient.GetDeviceId(DeviceName);
 
-                if (deviceid == null) return;
-                if (WaitForComplete.IsPresent)
-                {
-                    var jobstatusInfo = StorSimpleClient.DeleteDataContainer(deviceid, VolumeContainer.InstanceId);
-                    WriteObject(jobstatusInfo);
-                }
-                else
-                {
-                    var jobresult = StorSimpleClient.DeleteDataContainerAsync(deviceid, VolumeContainer.InstanceId);
-                    var msg =
-                            "Job submitted succesfully. Please use the command Get-AzureStorSimpleJob -InstanceId " +
-                            jobresult.JobId + " for tracking the job status";
-                        WriteObject(msg);
-                }
-            }
-            catch (CloudException cloudException)
-            {
-                StorSimpleClient.ThrowCloudExceptionDetails(cloudException);
-            }
+                                  if (deviceid == null) return;
+                                  if (WaitForComplete.IsPresent)
+                                  {
+                                      var jobstatusInfo = StorSimpleClient.DeleteDataContainer(deviceid, VolumeContainer.InstanceId);
+                                      WriteObject(jobstatusInfo);
+                                  }
+                                  else
+                                  {
+                                      var jobresult = StorSimpleClient.DeleteDataContainerAsync(deviceid, VolumeContainer.InstanceId);
+                                      WriteObject(ToAsyncJobMessage(jobresult, "delete"));
+                                  }
+                              }
+                              catch (CloudException cloudException)
+                              {
+                                  StorSimpleClient.ThrowCloudExceptionDetails(cloudException);
+                              }
+                          });
         }
     }
 }
